feat: Add topic and schema definition
This is a proposal for how topics and schemas should be defined in this repository.
Each topic is a yaml file in the `topics` directory. Each schema is a file in the
`schemas` directory. Topics can reference a schema. The example provided is a json
schema example, but this should be easily extendable for other types.

The example provided is for the `querylog` topic which is used by Snuba.