feat: Add basic scaffolding for Python library
This patch adds the scaffolding for a Python library under the
`python` directory. This allows us to easily import our schemas
and topic configurations into Python projects. Eventually, we want
to publish libraries in other languages as needed.

The API is very minimal, there is just one function:
`get_topic_configuration(topic)` takes a topic string and returns all of the topic config.
This includes 2 things: the topic creation config, and the schema if one is defined
for that topic. If the topic itself is not defined, it returns None.

This function is just a placeholder for now since we don't have any schemas defined here yet.