feat: Add topic and schema definition (#2)
This PR adds the basic structure for the schemas repository outlined in https://github.com/getsentry/rfcs/blob/main/text/0072-kafka-schema-registry.md. Each topic is a yaml file in the `topics` directory. Each schema is a file in the `schemas` directory. Example payloads can be added as files in the `examples` directory. Topics can reference schemas and examples.

The example provided is for the `querylog` topic which is used by Snuba.