feat: Remove unused fields from subscription results (#88)
Whilst creating this schema and integrating it into Snuba and Sentry we discovered there were some fields ("profile" and "trace_output") being published that were essentially internal Snuba data and are completely unused by Sentry in the results consumer. Let's remove them from the schema and example.

Since "additionalProperties" is true, this is a backward compatible change and old messages still in the pipeline will not fail validation.