Embed Topics/Schemas into Rust crate
When you ship a final binary using the Rust crate, that system obviously does not have all the static data available in `CARGO_MANIFEST_DIR` which is the crate path at compile time.

We thus embed all the topic metadata and schemas into the crate directly so it is available at runtime.