feat: Topic config
This repo should be the source of truth for creating topics needed
across various services with the correct configuration.
Topics can now specify topic creation config.

This also adds all of the commit log topics. This is needed
as those topics has specific creation config that is required,
and we want to have this in one place.