ref: Import errors schema from Snuba
The current errors schema from Relay is overly complicated. Let's use
the generated one from Snuba and see if it passes schema validation too.

Intentionally punting on a solution for keeping the schemas in sync --
we have tests in Snuba for that which currently fail, because Relay's
schema is too complex to be compared using json-schema-diff.