feat(spans): Add topic definition for buffered-segment topic (#237)
We're building a new pipeline to support performance issues with span streaming,
at a high level, it looks like this
Screenshot 2024-03-13 at 5 00 35 PM

the buffered-segment topic will expect to receive a list of spans. As of this
moment, these spans will have the exact same schema as ones we consume
from the spans kafka topic, the SpanEvent definition is just copied from
the snuba-span schema. we might deserialize the spans to drop fields before pushing to redis
as a memory optimization at which point the span schemas will diverge.