feat: Add required partition count to topic definition (#250)
There are a handful of topics at sentry that require partitions=1 and will not work properly with higher counts. Specifically the commit log topics.

Currently we only define partition counts per region in ops. This PR proposes adding the count to the topic definition where it required and enforcing this in ops (which imports this repo).