introduce standard segment settings for all commit log topics
very large segments that roll over infrequently are problematic on commit log
topics because the topics will roll over infrequently causing the topics to
get too large. this is problematic for post process forwarders specifically
which need to scan the entire commit log on startup.

these values are taken from the events and transactions commit log topics in sentry
production, we need to actually apply them across the board.