introduce standard segment settings for all commit log topics (#281)
very large segments are problematic on commit log topics because the topics will roll over infrequently causing the topics to get too large. this is specifically not good for post process forwarders which need to scan the entire commit log on startup. we are starting to see large backlogs on deployment of the issue platform PPF today.

these values are taken from the events and transactions commit log topics in sentry production, we need to actually apply them across the board.