feat(crons): Add mark_unknown clock task
This task will be triggered when we detect an anomaly in check-in
volume during a clock tick. When this happens we are unable to know that
all check-ins before that tick have been received and will need to mark
all in-progress monitors as resulting in a 'unknown' state.

Part of https://github.com/getsentry/sentry/issues/79328