feat: Add taskworker schema
The changes in getsentry/sentry#79742 are blocked because the
`task-worker` topic does not have schema metadata.

While I would like this topic to contain protobuf messages, we lack
tooling for that here at this time. As a temporary solution, I've added
JSON schema for the task-worker topic, and will work on adding protobuf
support separately.