feat(protobuf) Add support for protobuf message topics (#344)
For taskworkers (and likely other topics) in the future, we would like
to put protobuf serialized messages into the topic instead of
json/msgpack. Using protobufs will let us have one message schema for
both kafka, rpc and application logic.

In order to get access to the protobuf generated code, I've needed to
expand the dependencies of this library. I thought this was the simplest
to build and maintain solution.

Our protos package isn't compatible with python 3.9. 3.9 went end of
life mid 2022 and I don't think we have any applications still running
on 3.9