feat(uptime): Add optional fields to the `RequestInfo` schema to track more request detail
This adds timing and extra request detail to the uptime-results `RequestInfo` object. These fields are all optional, and can be null when provided.

Once we have this info, these fields should always be set, and if the information is not available then they should be null.

We can switch these fields over to being required once we're sending this info.