feat(uptime): Add optional fields to the `RequestInfo` schema to track more request detail (#416)
* feat(uptime): Add optional fields to the `RequestInfo` schema to track more request detail

This adds timing and extra request detail to the uptime-results `RequestInfo` object. These fields are all optional, and can be null when provided.

Once we have this info, these fields should always be set, and if the information is not available then they should be null.

We can switch these fields over to being required once we're sending this info.

* update schema after discussions

* style(lint): Auto commit lint changes

* redundant

* include url

* naming

* include total duration

* clarify ttfb

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>