fix(uptime): Refine new fields on the `uptime-results` schema
Due to how we defined some fields in the schema, the code wasn't generated correctly. Also taking the opportunity to skip allowing fields to nullable. Instead, if the field isn't available, it should just not be set, instead of being null. This aligns with sending data to EAP as well, since EAP doesn't support null attributes.